tests/requirements/base.txt
# requirements from setup.py
django-filer>=1.4.0
djangocms-picture>=2.1.0
djangocms-link>=2.2.1
django-polymorphic>=2.0.3
Pillow
html5lib>=0.999999999

# other requirements
coverage
django-app-helper
tox

# IMPORTANT: latest easy-thumbnails causes error since it returns with
# floating point, but the tests are not prepared for this and even the lib
easy-thumbnails

# In order to run skipped tests uncomment the next lines:
# -e git+ssh://git@github.com/divio/djangocms-transfer.git@master#egg=djangocms-transfer
# -e git+ssh://git@github.com/divio/djangocms-translations.git@master#egg=djangocms-translations
